Move `metamorph` package to `metafix` (#35)
