+++ org.metafacture.fix/src/main/java/org/metafacture/metafix/FixMethod.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.metafacture.metamorph;
+package org.metafacture.metafix;
 
 import org.metafacture.metamorph.maps.FileMap;
