+++ org.metafacture.fix/src/main/resources/flux-commands.properties
@@ -12,4 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-fix org.metafacture.metamorph.Metafix
+fix org.metafacture.metafix.Metafix
