+++ org.metafacture.fix/src/test/java/org/metafacture/metafix/InterpreterTest.java
@@ -1,8 +1,8 @@
-package org.metafacture.metafix.tests;
+package org.metafacture.metafix;
 
 import org.metafacture.metafix.fix.Fix;
 import org.metafacture.metafix.interpreter.FixInterpreter;
-import org.metafacture.metamorph.Metafix;
+import org.metafacture.metafix.tests.FixInjectorProvider;
 
 import com.google.inject.Inject;
 import org.eclipse.xtext.testing.InjectWith;
